fix: post request
The tests falied when using x-www-form-urlencoded as a header and then
creating `new URLSearchParams({query})` I don't know if this is using
the correct parameter so all the tests game a 400.

I switched out the header to use `application/sparql-query` and that
solved the issue.

I did have to modify one test to include the `:` prefix. Presumably this
is because there is implicit handling of namespaces on `GET` requests
but not `POST`